finish bot redo
